Folders, files, syntax main refactors
