+++ src/services/common.service.js
@@ -1,0 +1,11 @@
+export default {
+
+  isEmpty (param) {
+    return (param == null || param === '' || param === undefined)
+  },
+
+  isEmptyArray (param) {
+    return (param.length === 0)
+  }
+
+}
