+++ src/store/main.store.js
@@ -1,7 +1,8 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
-import books from './modules/books'
-import student from './modules/student'
+import books from './modules/book.store'
+import student from './modules/student.store'
+
 Vue.use(Vuex)
 
 export default new Vuex.Store({
